Drop codecs mod again, move it to media (but not pub)
